`ResponsesBuilder`: try to merge response with an existing one instead of overwriting
Before this commit, calling `ResponsesBuilder::response` with a status code that already has a response would cause the old response to be overwritten by the new one.

This commmit attempts to fix this behaviour slightly: in cases where a merge is possible, the responses will be merged. If a merge is not possible, we fall back to the original behaviour, i.e. overwriting the old response with the new one.

If a merge is possible, this improvement will merge their descriptions (by joining them with "\n\nOR\n\n"), examples and extensions.